+++ client/src/layout/MainLayout/Header/HeaderContent/sudoModal.jsx
@@ -28,6 +28,12 @@ const SudoModal = () => {
       try {
         setLoading(true);
         await API.auth.sudo(values);
+        
+        // if on /cosmos-ui/config-general refresh page because we need new infos
+        if (window.location.pathname.includes('/cosmos-ui/config-general')) {
+          window.location.reload();
+        }
+        
         setOpen(false);
         resetForm();
       } catch (error) {